# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HS Knowledge Base API",
    "version": "2.0.0",
    "summary": "Kho tri thức mở về mã HS, thuế và pháp lý xuất nhập khẩu Việt Nam",
...
    "license": {
      "name": "MIT (code) / CC BY-SA 4.0 (dữ liệu)",
      "url": "https://github.com/ozvietnam/hs-code-api/blob/main/NOTICE.md"
    },
    "contact": {
      "name": "HS Knowledge Base",
      "url": "https://github.com/ozvietnam/hs-code-api"
    }
  },
  "servers": [
    {
      "url": "https://hs-kb.uythacnhapkhau.com",
      "description": "Production"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "HS_API_TOKEN"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Thông điệp lỗi"
          },
          "code": {
            "type": "string",
            "description": "Mã lỗi ổn định (nếu có), vd INVALID_HS_CODE, FEEDBACK_NOT_PERSISTED"
          },
          "retryable": {
            "type": "boolean",
            "description": "true = lỗi tạm thời, gọi lại sau"
          },
          "detail": {
            "type": "string",
            "description": "Chi tiết kỹ thuật"
          }
        }
      },
      "AcftaForOrigin": {
        "type": "object",
        "description": "Mức ACFTA áp cho nước xuất xứ. eligible=false → áp MFN; null → biểu gốc có nhiều mức, tra dòng 10 số.",
        "properties": {
          "origin": {
            "type": "string",
            "description": "Mã ISO-2"
          },
          "eligible": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "rate": {
            "type": [
              "number",
              "null"
            ]
          },
          "higherThanMfn": {
            "type": "boolean",
            "description": "ACFTA cao hơn MFN — nên khai MFN"
          },
          "noteVi": {
            "type": "string",
            "description": "Giải thích để đọc cho người dùng"
          }
        }
      },
      "MissingFact": {
        "type": "object",
        "properties": {
          "attribute": {
            "type": "string",
            "description": "Tên khoá dùng trong facts"
          },
          "questionVi": {
            "type": "string",
            "description": "Câu hỏi để hỏi người dùng"
          },
          "type": {
            "type": "string",
            "enum": [
              "enum",
              "number"
            ]
          },
          "unit": {
            "type": "string",
            "description": "Đơn vị khi type=number"
          },
          "optionsVi": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "index": {
                  "type": "integer"
                },
                "value": {
                  "type": "string"
                },
                "labelVi": {
                  "type": "string"
                }
              }
            },
            "description": "Lựa chọn hợp lệ. Trả lời bằng value, index hoặc labelVi đều được."
          }
        }
      },
      "GirDetermination": {
        "type": "object",
        "properties": {
          "rule": {
            "type": "string",
            "description": "Quy tắc, vd \"GIR 6\""
          },
          "basis": {
            "type": "string",
            "enum": [
              "RULE_TABLE",
              "DETERMINISTIC",
              "HEURISTIC",
              "LLM_ASSERTED"
            ]
          },
          "reasonVi": {
            "type": "string",
            "description": "Lý do"
          },
          "evidence": {
            "type": "object"
          },
          "source": {
            "type": "string",
            "description": "Nguồn"
          }
        }
      },
      "Suggestion": {
        "type": "object",
        "properties": {
          "hsCode": {
            "type": "string",
            "description": "Mã 8 số — luôn có trong biểu thuế"
          },
          "nameVi": {
            "type": "string",
            "description": "Tên trong biểu thuế"
          },
          "confidence": {
            "type": [
              "number",
              "null"
            ],
            "description": "Điểm mô hình tự khai, CHƯA hiệu chuẩn — không phải xác suất. null ở chế độ deterministic."
          },
          "reasoning": {
            "type": "string",
            "description": "Lý do; số hiệu văn bản chưa kiểm chứng được gắn \"[chưa kiểm chứng]\""
          },
          "unverifiedCitations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "productExamples": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Tên hàng THẬT từ tờ khai (mã Loại khác)"
          },
          "productExamplesGenerated": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Câu máy sinh, chưa kiểm chứng"
          },
          "addedByResidualGuard": {
            "type": "boolean",
            "description": "Mã Loại khác do hệ thống chèn thêm để cân nhắc"
          },
          "taxAcftaChina": {
            "$ref": "#/components/schemas/AcftaForOrigin"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/search": {
      "get": {
        "operationId": "search",
        "summary": "Tìm mã HS theo từ khoá hoặc mã một phần",
        "description": "Tra được cả tên hàng theo cách người đi khai gọi, không chỉ lời văn biểu thuế. Mỗi kết quả có thể kèm: `precedent` (cụm từ này đã được khai bao nhiêu lần trong tờ khai đã thông quan, kèm mức tập trung), `tradeTerm` (mục từ điển tên thương mại đã khớp — ĐỌC `appliesWhenVi` trước khi dùng, vì mã đúng thường phụ thuộc khổ, dạng hoặc thành phần), `breadcrumb` (Phần → Chương → Nhóm → Phân nhóm; với mã tên là \"Loại khác\" thì `scopeVi` mới là phạm vi thật và `excludesVi` là các mã anh em bị loại trừ), và `vatReduction` (`eligible: false` nghĩa là KHÔNG được giảm VAT theo NĐ 174/2025 — khai 8% sẽ bị truy thu). Cấp response còn có `avoidedCodes` (mã đã bị loại khỏi kết quả vì là bẫy nhầm lẫn, kèm lý do), `clarifyingQuestionsVi` (dữ kiện còn thiếu để chốt mã), `formWarning` (hỏi nguyên liệu nhưng tiền lệ là thành phẩm) và `tradeTermNotApplied`. Khi có `avoidedCodes` hoặc `clarifyingQuestionsVi`, hãy nói lại cho người dùng thay vì nuốt đi. Tên gọi chỉ đưa tới nhóm 4 số; lá 8 số do BẢNG QUYẾT ĐỊNH theo thuộc tính chốt: `decisions[]` (mỗi nhóm có bảng: `status` RESOLVED/INSUFFICIENT, `hsCode` + `ruleId` + `reasonVi` + `source` khi chốt được, `missingFacts[]` là thuộc tính còn thiếu kèm `questionVi`, `basis` RULE_TABLE chỉ khi bảng đã verified). Kết quả do bảng chốt có `decision` và đứng đầu. Trả lời câu hỏi bằng `facts` để chốt.",
        "tags": [
          "Tra cứu"
        ],
        "security": [],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "description": "Từ khoá tiếng Việt hoặc mã HS một phần",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "Số kết quả tối đa (mặc định 10)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "facts",
            "in": "query",
            "required": false,
            "description": "JSON object dữ kiện tường minh theo tên thuộc tính trong `missingFacts`, vd {\"thicknessMm\":2,\"form\":\"coil\"}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi máy chủ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/suggest": {
      "post": {
        "operationId": "suggest",
        "summary": "Gợi ý mã HS bằng AI, kèm audit trail GIR có căn cứ",
        "description": "Body: `{\"description\": \"tên hàng\"}`. Trả `suggestions[]`, `girRulesApplied[]` (mỗi mục có `basis`: RULE_TABLE/DETERMINISTIC là căn cứ chắc, HEURISTIC/LLM_ASSERTED phải kiểm chứng lại), `girDisclaimer`, `rankingSignals[]` (kỹ thuật, không có giá trị pháp lý) và `chapterGuidance[]`. Kèm `decisions[]` (bảng quyết định theo thuộc tính cho các nhóm trong top gợi ý) và `missingFacts[]` — thuộc tính còn thiếu để chốt lá 8 số; ERP/người dùng trả lời bằng body `facts: {...}` rồi gọi lại. Bảng đã verified chốt được lá thì lá đó lên đầu với `decidedByTable`; chưa verified chỉ tư vấn. Kèm `confusionAlerts[]` từ từ điển mâu thuẫn HS (xem /api/confusion-pairs). Mở công khai, không cần token (mỗi lượt gọi tốn chi phí LLM nên hãy gọi có chủ đích). ĐỌC `status` + `nextAction` TRƯỚC: NEED_FACTS → hỏi người dùng rồi gọi lại với `facts`; NEEDS_EXPERT → AI không chạy được, cần chuyên viên; REVIEW → người dùng chọn/xác nhận. Không có trạng thái tự chốt: `confidence` chưa hiệu chuẩn.",
        "tags": [
          "AI"
        ],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string",
                    "description": "Tên hàng + chất liệu + công dụng + thông số (≥ 3 ký tự)",
                    "minLength": 3
                  },
                  "facts": {
                    "type": "object",
                    "additionalProperties": true,
                    "description": "Trả lời missingFacts: {<attribute>: <value | index | labelVi | số>}"
                  },
                  "options": {
                    "type": "object",
                    "properties": {
                      "topReranked": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 5,
                        "default": 3
                      },
                      "topCandidates": {
                        "type": "integer",
                        "minimum": 3,
                        "maximum": 20,
                        "default": 10
                      }
                    }
                  },
                  "items": {
                    "type": "array",
                    "maxItems": 20,
                    "description": "Chế độ batch: [{id, description}] — thay cho description",
                    "items": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "NO_CANDIDATES",
                        "NEED_FACTS",
                        "NEEDS_EXPERT",
                        "RESOLVED_BY_TABLE",
                        "REVIEW"
                      ],
                      "description": "Đọc trước tiên"
                    },
                    "nextAction": {
                      "type": "object",
                      "description": "Việc agent cần làm tiếp",
                      "properties": {
                        "type": {
                          "type": "string",
                          "enum": [
                            "REPHRASE",
                            "ASK_USER",
                            "HUMAN_REVIEW",
                            "USER_CONFIRM"
                          ]
                        },
                        "questions": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/MissingFact"
                          }
                        },
                        "then": {
                          "type": "object",
                          "description": "Lời gọi tiếp theo (endpoint + body mẫu)"
                        },
                        "optionsHs": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "reasonVi": {
                          "type": "string"
                        }
                      }
                    },
                    "suggestions": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Suggestion"
                      }
                    },
                    "engine": {
                      "type": "string",
                      "enum": [
                        "llm",
                        "deterministic"
                      ]
                    },
                    "degraded": {
                      "type": "boolean"
                    },
                    "llmRejectedCodes": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "hsCode": {
                            "type": "string"
                          },
                          "reason": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "missingFacts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MissingFact"
                      }
                    },
                    "rejectedFacts": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "decisions": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "girRulesApplied": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/GirDetermination"
                      }
                    },
                    "residualAdvisory": {
                      "type": "object"
                    },
                    "antiPatternWarnings": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "confusionWarning": {
                      "type": [
                        "object",
                        "null"
                      ]
                    },
                    "confusionAlerts": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      },
                      "description": "Từ điển mâu thuẫn HS: DN hay khai A, Hải quan hay ấn định B"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Thiếu description (≥ 3 ký tự) hoặc JSON sai",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi máy chủ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/classify": {
      "post": {
        "operationId": "classify",
        "summary": "Phân loại có cây quyết định + bảng phân giải cụm mã dễ nhầm",
        "description": "Đọc `status` + `nextAction` TRƯỚC (cùng bảng với /api/suggest: REVIEW, NEED_FACTS, NEEDS_EXPERT [LOW_CONFIDENCE, FEATURE_CONFLICT, ENGINE_TIMEOUT, LLM_REJECTED], RESOLVED_BY_TABLE, NO_CANDIDATES). Nhận nguyên tiêu đề Taobao tiếng Trung (`queryUnderstanding`). Trả `results[]` (mã 8 hoặc 6 số), `missingStructured[]` (trường bắt buộc còn thiếu theo nhóm 4 số, câu hỏi VI + ZH), `originAssessment`, `girRulesApplied[]` (có `basis`), `llmRejectedCodes` khi AI trả mã không có trong biểu thuế.",
        "tags": [
          "AI"
        ],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "tenHang"
                ],
                "properties": {
                  "tenHang": {
                    "type": "string",
                    "description": "Tên hàng (≥ 2 ký tự)"
                  },
                  "chatLieu": {
                    "type": "string",
                    "description": "Chất liệu"
                  },
                  "congDung": {
                    "type": "string",
                    "description": "Công dụng"
                  },
                  "chucNang": {
                    "type": "string",
                    "description": "Chức năng"
                  },
                  "specs": {
                    "type": "string",
                    "description": "Thông số"
                  },
                  "nameZh": {
                    "type": "string",
                    "description": "Tên tiếng Trung"
                  },
                  "tier": {
                    "type": "string",
                    "enum": [
                      "standard",
                      "premium"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Thiếu tenHang",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Lỗi máy chủ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "Lỗi phân loại",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}